requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dj-database-url==2.3.0
Django==5.2
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.18.0
fsspec==2025.3.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
ollama==0.4.7
openai==1.76.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
psycopg2==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
